Rename “Stages” to “Levels”
In video games, the terms “stage” and “level” mean the same thing, so
either term works. I think that level is the superior term in this
situation. “Stage” is a metaphor based on theater. Godot is named after
a play and already uses the term “scene” as a metaphor. In theater,
scenes are preformed on a stage. Before this change, in this game, each
Stage was a scene (which isn’t how it works in theater).