Correct note about where SPDX data is in scenes
For scenes that inherit from other scenes, I can’t necessarily put SPDX
metadata on the root Node. This is because Godot is efficient: if the
metadata for a Node in the parent scene matches the metadata for that
Node in the child scene, then it won’t store that data twice.